online_store/requirements.txt
Django>=4.1.7,<5.0
Pillow>=9.4.0,<10.0
pyyaml>=6.0,<7.0
django-safedelete>=1.3.1,<2.0